# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "DriveDaysOfWeekEnum": {
        "enum": [
          "Mon",
          "Tue",
          "Wed",
          "Thu",
          "Fri"
        ],
        "title": "DriveDaysOfWeekEnum",
        "type": "string"
      },
      "DriverAssignmentStatusEnum": {
        "description": "Whether the routes in a RouteGroup have drivers assigned (read from\nRoute.driver_id).",
        "enum": [
          "Assigned",
          "Unassigned"
        ],
        "title": "DriverAssignmentStatusEnum",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaginatedResponse_RouteWithDateRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RouteWithDateRead"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "total_pages": {
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "title": "PaginatedResponse[RouteWithDateRead]",
        "type": "object"
      },
...
      "RouteGroupCreate": {
        "description": "Create request model",
        "properties": {
          "drive_date": {
            "format": "date-time",
            "title": "Drive Date",
            "type": "string"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          }
        },
        "required": [
          "name",
          "drive_date"
        ],
        "title": "RouteGroupCreate",
        "type": "object"
      },
      "RouteGroupRead": {
        "description": "Read response model",
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "delivery_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Type"
          },
          "drive_date": {
            "format": "date-time",
            "title": "Drive Date",
            "type": "string"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "num_boxes": {
            "default": 0,
            "title": "Num Boxes",
            "type": "integer"
          },
          "num_drivers_assigned": {
            "default": 0,
            "title": "Num Drivers Assigned",
            "type": "integer"
          },
          "num_locations": {
            "default": 0,
            "title": "Num Locations",
            "type": "integer"
          },
          "num_routes": {
            "title": "Num Routes",
            "type": "integer"
          },
          "route_group_id": {
            "format": "uuid",
            "title": "Route Group Id",
            "type": "string"
          },
          "routes": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RouteReadSummary"
            },
            "title": "Routes",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/RouteStatusEnum"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "name",
          "drive_date",
          "route_group_id",
          "num_routes",
          "status"
        ],
        "title": "RouteGroupRead",
        "type": "object"
      },
      "RouteGroupUpdate": {
        "description": "Update request model - all optional",
        "properties": {
          "drive_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drive Date"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "title": "RouteGroupUpdate",
        "type": "object"
      },
      "RoutePatchRequest": {
        "description": "Request body for PATCH /routes/{route_id}.\n\nAll fields are optional - only provided fields will be updated.\nIf location_ids is provided, the route stops will be fully replaced\nand the routing algorithm will be re-run to update polyline + mileage.",
        "properties": {
          "driver_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Id"
          },
          "location_ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Ids"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "start_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          }
        },
        "title": "RoutePatchRequest",
        "type": "object"
      },
...
      "RouteReadSummary": {
        "description": "Lightweight route info returned inside a route group response",
        "properties": {
          "length": {
            "default": 0,
            "title": "Length",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "route_id": {
            "format": "uuid",
            "title": "Route Id",
            "type": "string"
          }
        },
        "required": [
          "route_id",
          "name"
        ],
        "title": "RouteReadSummary",
        "type": "object"
      },
      "RouteStatusEnum": {
        "enum": [
          "Upcoming",
          "Completed",
          "Archived"
        ],
        "title": "RouteStatusEnum",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend API for the Food4Kids application",
    "title": "Food4Kids API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/route-groups": {
      "get": {
        "description": "Retrieve all route groups, optionally filtered by date range, weekday, delivery type, route status, and driver assignment status.\nCan include associated routes in the response.",
        "operationId": "get_route_groups",
        "parameters": [
          {
            "description": "Filter route groups from this date",
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter route groups from this date",
              "title": "Start Date"
            }
          },
          {
            "description": "Filter route groups until this date",
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter route groups until this date",
              "title": "End Date"
            }
          },
          {
            "description": "Filter by one or more weekdays",
            "in": "query",
            "name": "weekday",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/DriveDaysOfWeekEnum"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by one or more weekdays",
              "title": "Weekday"
            }
          },
          {
            "description": "Filter by one or more delivery types",
            "in": "query",
            "name": "delivery_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by one or more delivery types",
              "title": "Delivery Type"
            }
          },
          {
            "description": "Filter by one or more route statuses",
            "in": "query",
            "name": "route_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/RouteStatusEnum"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by one or more route statuses",
              "title": "Route Status"
            }
          },
          {
            "description": "Filter by one or more driver assignment statuses",
            "in": "query",
            "name": "driver_assignment_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/DriverAssignmentStatusEnum"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by one or more driver assignment statuses",
              "title": "Driver Assignment Status"
            }
          },
          {
            "description": "Include routes in the response",
            "in": "query",
            "name": "include_routes",
            "required": false,
            "schema": {
              "default": false,
              "description": "Include routes in the response",
              "title": "Include Routes",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RouteGroupRead"
                  },
                  "title": "Response Get Route Groups",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Route Groups",
        "tags": [
          "route-groups"
        ]
      },
      "post": {
        "description": "Create a new route group",
        "operationId": "create_route_group",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteGroupCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteGroupRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Route Group",
        "tags": [
          "route-groups"
        ]
      }
    },
    "/route-groups/{route_group_id}": {
      "delete": {
        "description": "Delete a route group and all its route group memberships",
        "operationId": "delete_route_group",
        "parameters": [
          {
            "in": "path",
            "name": "route_group_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Group Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Route Group",
        "tags": [
          "route-groups"
        ]
      },
      "patch": {
        "description": "Update an existing route group",
        "operationId": "update_route_group",
        "parameters": [
          {
            "in": "path",
            "name": "route_group_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Group Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteGroupUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteGroupRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Route Group",
        "tags": [
          "route-groups"
        ]
      }
    },
    "/route-groups/{route_group_id}/duplicate": {
      "post": {
        "description": "Duplicate a route group and its routes/stops for a new planning cycle.",
        "operationId": "duplicate_route_group",
        "parameters": [
          {
            "in": "path",
            "name": "route_group_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Group Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteGroupRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Duplicate Route Group",
        "tags": [
          "route-groups"
        ]
      }
    },
    "/routes": {
      "get": {
        "description": "Get routes with pagination and optional filtering for unassigned routes and date range.\nReturns routes with their drive dates - routes can appear multiple times for different dates.\nWhen unassigned_only is False, returns all routes (no assignment filter).\nWhen unassigned_only is True, returns only routes that are unassigned for the given route group.\n\nRequires a driver or admin caller.\nAdmins may scope to any driver via driver_id (or omit it for all routes).\nDrivers are always scoped to their own routes: omitting driver_id returns\ntheir own routes, and requesting another driver's is rejected.",
        "operationId": "get_routes",
        "parameters": [
          {
            "description": "If true, only return unassigned routes. If false, return all routes regardless of assignment status.",
            "in": "query",
            "name": "unassigned_only",
            "required": false,
            "schema": {
              "default": false,
              "description": "If true, only return unassigned routes. If false, return all routes regardless of assignment status.",
              "title": "Unassigned Only",
              "type": "boolean"
            }
          },
          {
            "description": "Filter route groups from this date",
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "description": "Filter route groups from this date",
              "title": "Start Date",
              "type": "string"
            }
          },
          {
            "description": "Filter route groups until this date",
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "description": "Filter route groups until this date",
              "title": "End Date",
              "type": "string"
            }
          },
          {
            "description": "Order by drive_date: 'asc' (default, oldest-first) for the upcoming feed, 'desc' (most-recent-first) for the past feed.",
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "asc",
              "description": "Order by drive_date: 'asc' (default, oldest-first) for the upcoming feed, 'desc' (most-recent-first) for the past feed.",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Order",
              "type": "string"
            }
          },
          {
            "description": "Page number (1-indexed)",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Page number (1-indexed)",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Number of items per page",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Number of items per page",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "description": "Filter routes by assigned driver. Admins may pass any driver_id, or omit it to get all routes. Drivers are always scoped to themselves: omitting it returns their own routes, and passing another driver's id is rejected with 403.",
            "in": "query",
            "name": "driver_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter routes by assigned driver. Admins may pass any driver_id, or omit it to get all routes. Drivers are always scoped to themselves: omitting it returns their own routes, and passing another driver's id is rejected with 403.",
              "title": "Driver Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_RouteWithDateRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Routes",
        "tags": [
          "routes"
        ]
      }
    },
    "/routes/{route_id}": {
      "delete": {
        "description": "Delete a route by its unique identifier.\n\nParameters:\n    route_id (UUID): The unique identifier of the route to delete.\n    session (AsyncSession): The database session dependency.\n    _ (bool): Indicates that the user is authenticated as a driver (injected by dependency).\n\nReturns:\n    None. Responds with HTTP 204 No Content on successful deletion.",
        "operationId": "delete_route",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Route",
        "tags": [
          "routes"
        ]
      },
      "get": {
        "description": "Get a route by its unique identifier, with its ordered stops embedded.\n\nEach stop carries its sequence #, address, contact name, phone (+ secondary),\nbox count, and a note_chain_id reference. Stops are sourced live from\nLocation for upcoming routes and from frozen route_stop_snapshots for past\nroutes. Notes are not embedded: fetch them via GET /note-chains/{id}/notes.\n\nParameters:\n    route_id (UUID): The unique identifier of the route to GET.\n    session (AsyncSession): The database session dependency.\n\nReturns:\n    The route with its ordered stops. Responds with HTTP 200 OK on success.",
        "operationId": "get_route",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteDetailRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Route",
        "tags": [
          "routes"
        ]
      },
      "patch": {
        "description": "Update a route's metadata (name, notes) and/or its stop order/locations.\n\nIf location_ids is provided in the request body:\n- The route's stops are fully replaced with the new ordered list.\n- The routing algorithm is re-run to compute the new polyline and mileage.\n- This will affect ALL route groups that share this route.\n\nIf only name/notes are provided, stops and mileage are left unchanged.",
        "operationId": "update_route",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutePatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Route",
        "tags": [
          "routes"
        ]
      }
    },
...
  }
}